UserExt: add objects that can be used in place of [], {}, or null
Normally key-value pairs with the value being {}, [], or null get
scrubbed out. A whitelist decided when to keep them.

This creates a better system for intentionally putting these in
the glTF file that can also be used by user extensions. Just put
gltf2_io.EmptyDict in place of where you want, eg. {}, and it
will work.