Export: handle zero-weight verts by inserting a new bone
When skinning in Blender, verts that aren't assigned to bones act
as if they weren't skinned. In glTF all verts must be assigned to
bones. So to handle these verts, we insert an untransformed
"neutral bone" into the armature for a mesh that has unassigned
verts and treat those verts as if they were assigned to the neutral
bone.

The neutral bone is a root bone with no TRS, so skinning with it
should be the same as not skinning at all.

The neutral bone should only be inserted if necessary. If there
are no verts that are unassigned, there should be no change.

This is kinda awkward to do. We do it with a second pass after
a scene has been gathered. extract_primitives marks what skins
need the neural bone, and the second pass finds all those skins,
their arma nodes, and inserts the new bone.